Add custom fields for template inheritance
Hi. We did some change to T4 template which could be useful.
The change introduces custom properties to the template items, so they could be explicitly told to inherit specifc class/interface.

This could be useful if there are TDS templates which inherits other TDS templates which are not part of the project.